requirements.txt
streamlit
pillow
numpy
opencv-python
torch
torchvision
scikit-learn
matplotlib
seaborn
torchsummary
pandas
joblib
streamlit_option_menu
hydralit_components